Add 'beep' bin.
This is just an easy way to ring the terminal bell. Because of the
settings elsewhere (in ~/.Xdefaults and ~/.tmux.conf) that propagate
bells and convert them to window manager urgent hints, this is a cheap
way to notify yourself when a command is done (*after* you've already
started it).

Type 'beep' while a long compilation or whatever is happening; your
input should be buffered, and 'beep' will run when the previous command
completes.

That's the idea, anyway. I've tried this exactly once.