Revamp method for ignoring files
Use `deploy.ignorefiles` as a list of basenames to ignore. It's not the
most flexible in that we can't ignore file foo in the top level
directory but allow it in subdirs -- unless it appears in a subdir of
sometihng we copy wholesale.

But t)e mechanics are very simple. Finally!