ssh-agent: use umask, more comments
Using umask avoids a potential race where someone could open a
filehandle to the env file before the chmod runs. (This is highly
theoretical, but no reason to allow it if we have a simple way to avoid
it.)

Added comments to attempt to do a better job of explaining this thing.