[sway] autorun dbus-update-activation-environment
For my setup, which is using GDM to log in to Sway (a stock Fedora
install with Sway installed; NOT the Sway Fedora spin), this seems to
get everything up and running so I can unlock the keychain.

Things like 1Password are able to store their temp 2FA tokens in the
keychain now, and I can open `seahorse` and unlock the `Login` keychain.

I didn't have to use a PAM module, and I don't want to since that's not
the stock Fedora setup.

I'm sure this will break again in the future in a way that will force me
to fully understand each component...